move nav items
